Added malloc checks in most relevant places & fixed segfault bug
Specifically, ran into an issue with windows that were too large
w/ disparate count rates (i.e. much larger sync rate than count rate
on relevant channels) that I think I've resolved but need to sanity
check tomorrow

+++ src/ttd_crosscorr2.h
@@ -13,8 +13,8 @@ typedef struct {
   ttd_t bin_time;
   ttd_t window_time;
 
-  int num_bins;
-  int center_bin;
+  int64_t num_bins;
+  int64_t center_bin;
 
   int64_t total;
